requirements.txt
sqlalchemy
python-dotenv
aiogram
asyncio
alembic
asyncpg
httpx